+++ google/iam.go
@@ -1,0 +1,96 @@
+package google
+
+import (
+	"fmt"
+	"github.com/hashicorp/terraform/helper/schema"
+	"google.golang.org/api/cloudresourcemanager/v1"
+	"log"
+	"time"
+)
+
+type ResourceIamUpdater interface {
+	GetResourceIamPolicy() (*cloudresourcemanager.Policy, error)
+	SetProjectIamPolicy(policy *cloudresourcemanager.Policy) error
+	GetMutexKey() string
+	GetResourceId() string
+	DescribeResource() string
+}
+
+type newResourceIamUpdaterFunc func(d *schema.ResourceData, config *Config) (ResourceIamUpdater, error)
+type iamPolicyModifyFunc func(p *cloudresourcemanager.Policy) error
+
+func iamPolicyReadModifyWrite(updater ResourceIamUpdater, modify iamPolicyModifyFunc) error {
+	mutexKey := updater.GetMutexKey()
+	mutexKV.Lock(mutexKey)
+	defer mutexKV.Unlock(mutexKey)
+
+	for {
+		backoff := time.Second
+		log.Printf("[DEBUG]: Retrieving policy for %s\n", updater.DescribeResource())
+		p, err := updater.GetResourceIamPolicy()
+		if err != nil {
+			return err
+		}
+		log.Printf("[DEBUG]: Retrieved policy for %s: %+v\n", updater.DescribeResource(), p)
+
+		err = modify(p)
+		if err != nil {
+			return err
+		}
+
+		log.Printf("[DEBUG]: Setting policy for %s to %+v\n", updater.DescribeResource(), p)
+		err = updater.SetProjectIamPolicy(p)
+		if err == nil {
+			break
+		}
+		if isConflictError(err) {
+			log.Printf("[DEBUG]: Concurrent policy changes, restarting read-modify-write after %s\n", backoff)
+			time.Sleep(backoff)
+			backoff = backoff * 2
+			if backoff > 30*time.Second {
+				return fmt.Errorf("Error applying IAM policy to %s: too many concurrent policy changes.\n", updater.DescribeResource())
+			}
+			continue
+		}
+		return fmt.Errorf("Error applying IAM policy for %s: %v", updater.DescribeResource(), err)
+	}
+	log.Printf("[DEBUG]: Set policy for %s", updater.DescribeResource())
+	return nil
+}
+
+// Merge multiple Bindings such that Bindings with the same Role result in
+// a single Binding with combined Members
+func mergeBindings(bindings []*cloudresourcemanager.Binding) []*cloudresourcemanager.Binding {
+	bm := rolesToMembersMap(bindings)
+	rb := make([]*cloudresourcemanager.Binding, 0)
+
+	for role, members := range bm {
+		var b cloudresourcemanager.Binding
+		b.Role = role
+		b.Members = make([]string, 0)
+		for m := range members {
+			b.Members = append(b.Members, m)
+		}
+		rb = append(rb, &b)
+	}
+
+	return rb
+}
+
+// Map a role to a map of members, allowing easy merging of multiple bindings.
+func rolesToMembersMap(bindings []*cloudresourcemanager.Binding) map[string]map[string]bool {
+	bm := make(map[string]map[string]bool)
+	// Get each binding
+	for _, b := range bindings {
+		// Initialize members map
+		if _, ok := bm[b.Role]; !ok {
+			bm[b.Role] = make(map[string]bool)
+		}
+		// Get each member (user/principal) for the binding
+		for _, m := range b.Members {
+			// Add the member
+			bm[b.Role][m] = true
+		}
+	}
+	return bm
+}
