+++ google/provider.go
@@ -132,8 +132,8 @@ func Provider() terraform.ResourceProvider {
 			"google_organization_policy":                   resourceGoogleOrganizationPolicy(),
 			"google_project":                               resourceGoogleProject(),
 			"google_project_iam_policy":                    resourceGoogleProjectIamPolicy(),
-			"google_project_iam_binding":                   resourceGoogleProjectIamBinding(),
-			"google_project_iam_member":                    resourceGoogleProjectIamMember(),
+			"google_project_iam_binding":                   ResourceIamBinding(IamProjectSchema, NewProjectIamUpdater),
+			"google_project_iam_member":                    resourceIamMember(IamProjectSchema, NewProjectIamUpdater),
 			"google_project_service":                       resourceGoogleProjectService(),
 			"google_project_iam_custom_role":               resourceGoogleProjectIamCustomRole(),
 			"google_project_services":                      resourceGoogleProjectServices(),
