+++ google/resource_google_project_iam_member_test.go
@@ -45,6 +45,8 @@ func TestAccGoogleProjectIamMember_multiple(t *testing.T) {
 	t.Parallel()
 
 	org := getTestOrgFromEnv(t)
+	skipIfEnvNotSet(t, "GOOGLE_ORG")
+
 	pid := "terraform-" + acctest.RandString(10)
 	resource.Test(t, resource.TestCase{
 		PreCheck:  func() { testAccPreCheck(t) },
@@ -86,6 +88,8 @@ func TestAccGoogleProjectIamMember_remove(t *testing.T) {
 	t.Parallel()
 
 	org := getTestOrgFromEnv(t)
+	skipIfEnvNotSet(t, "GOOGLE_ORG")
+
 	pid := "terraform-" + acctest.RandString(10)
 	resource.Test(t, resource.TestCase{
 		PreCheck:  func() { testAccPreCheck(t) },
